+++ R/dplyr_verbs_helpers.R
@@ -182,11 +182,19 @@ select_rename <- function(.eeg_lst, select = TRUE, ...) {
       # TODO: reimplement all the following directly in data table, or maybe with purrr
       old_channels <- events_tbl(.eeg_lst)$.channel
       new_channels <- all_vars[all_vars %in% old_channels] %>% sort()
-      # removes the old channels that do not exist anymore (needed for e.g., select(ZZ=X))
+      # removes the old channels that do not exist anymore 
+      ## (needed for e.g., select(ZZ=X))
+      ##
+      rem_pos <- which(!is.na(old_channels) & !old_channels %in% unname(new_channels))
       old_channels[!old_channels %in% new_channels] <- NA
+     ## old_channels <- old_channels[old_channels %in% new_channels]
+      
       events_tbl(.eeg_lst)$.channel <- old_channels %>%
         factor(labels = names(new_channels)) %>%
         as.character()
+      if(length(rem_pos)>0){
+        events_tbl(.eeg_lst) <- events_tbl(.eeg_lst)[-rem_pos]
+        }
     }
   }